creates new managers folder and renames files from service to manager and updates all dependencies
