+++ app/Http/Controllers/Legacy/Two/PostsController.php
@@ -2,7 +2,7 @@
 
 namespace Rogue\Http\Controllers\Legacy\Two;
 
-use Rogue\Services\Legacy\Two\PostService;
+use Rogue\Managers\Legacy\Two\PostManager;
 use Rogue\Http\Controllers\Traits\PostRequests;
 use Rogue\Repositories\Legacy\Two\SignupRepository;
 use Rogue\Http\Transformers\Legacy\Two\PostTransformer;
@@ -12,9 +12,9 @@ class PostsController extends ApiController
     use PostRequests;
 
     /**
-     * The post service instance.
+     * The post manager instance.
      *
-     * @var \Rogue\Services\Legacy\Two\PostService
+     * @var \Rogue\Managers\Legacy\Two\PostManager
      */
     protected $posts;
 
@@ -36,7 +36,7 @@ class PostsController extends ApiController
      * @param  PostContract  $posts
      * @return void
      */
-    public function __construct(PostService $posts, SignupRepository $signups)
+    public function __construct(PostManager $posts, SignupRepository $signups)
     {
         $this->posts = $posts;
         $this->signups = $signups;
