+++ app/Managers/Legacy/Two/SignupManager.php
@@ -1,18 +1,17 @@
 <?php
 
-namespace Rogue\Services;
+namespace Rogue\Managers\Legacy\Two;
 
 use Rogue\Jobs\SendSignupToQuasar;
 use Rogue\Jobs\SendSignupToCustomerIo;
-use Rogue\Repositories\SignupRepository;
-use Rogue\Jobs\SendDeletedSignupToQuasar;
+use Rogue\Repositories\Legacy\Two\SignupRepository;
 
-class SignupService
+class SignupManager
 {
     /*
      * SignupRepository Instance
      *
-     * @var Rogue\Repositories\SignupRepository;
+     * @var Rogue\Repositories\Legacy\Two\SignupRepository;
      */
     protected $signup;
 
@@ -27,16 +26,15 @@ class SignupService
         $this->signup = $signup;
     }
 
-    /**
+    /*
      * Handles all business logic around creating signups.
      *
      * @param array $data
-     * @param string $northstarId
      * @return Illuminate\Database\Eloquent\Model $model
      */
-    public function create($data, $northstarId)
+    public function create($data)
     {
-        $signup = $this->signup->create($data, $northstarId);
+        $signup = $this->signup->create($data);
 
         // Send to Blink unless 'dont_send_to_blink' is TRUE
         $should_send_to_blink = ! (array_key_exists('dont_send_to_blink', $data) && $data['dont_send_to_blink']);
@@ -50,51 +48,9 @@ class SignupService
         SendSignupToQuasar::dispatch($signup);
 
         // Log that a signup was created.
-        info('signup_created', ['id' => $signup->id]);
+        info('signup_created', ['id' => $signup->id, 'northstar_id' => $signup->northstar_id]);
 
         return $signup;
-    }
-
-    /**
-     * Handles all business logic around updating signups.
-     *
-     * @param Rogue\Models\Signup $signup
-     * @param array $data
-     * @return Rogue\Models\Signup $model
-     */
-    public function update($signup, $data)
-    {
-        $signup = $this->signup->update($signup, $data);
-
-        // Dispatch job to send signup to Quasar
-        SendSignupToQuasar::dispatch($signup);
-
-        // Log that a signup was updated.
-        info('signup_updated', ['id' => $signup->id]);
-
-        return $signup;
-    }
-
-    /**
-     * Handle all business logic around deleting a signup.
-     *
-     * @param int $signupId
-     * @return bool
-     */
-    public function destroy($signupId)
-    {
-        $trashed = $this->signup->destroy($signupId);
-
-        if ($trashed) {
-            info('signup_deleted', [
-                'id' => $signupId,
-            ]);
-
-            // Dispatch job to send post to Quasar
-            SendDeletedSignupToQuasar::dispatch($signupId);
-        }
-
-        return $trashed;
     }
 
     /*
