+++ app/Services/Fastly.php
