+++ app/Services/Registrar.php
